move sessions tests to own file
