+++ test/server-test.js
@@ -26,34 +26,4 @@ describe('Server', () => {
   it('should exist', () => {
     assert(app)
   })
-
-  describe('routes', () => {
-    beforeEach(() => {
-      return knex.migrate.rollback()
-      .then(() => knex.migrate.latest())
-      .then(() => knex.seed.run())
-    })
-
-    afterEach(() => {
-      return knex.migrate.rollback()
-    })
-
-    describe('POST /api/v1/sessions', () => {
-      it('logs a user in', (done) => {
-        let user_info = {username: 'johnyboi', password: 'pass'}
-
-        this.request.post('/api/v1/sessions', {form: user_info}, (err, res) => {
-          if (err) { done(err) }
-
-          let user_response = JSON.parse(res.body)
-
-          assert.equal(res.statusCode, 200)
-          assert.hasAllKeys(user_response, ['id','role'])
-          assert.equal(user_response.id, 1)
-          assert.equal(user_response.role, 'instructor')
-          done()
-        })
-      })
-    })
-  })
 })
